Updated package name and removed unneeded wrapper functions
